requirements.txt
numpy
pandas
psycopg2-binary
python-dateutil
python-dotenv
pytz
six
